extract queues
